# api/openapi.yaml
openapi: 3.1.0
info:
  title: AI Hackathon 2026 智能评审Agent系统 API
  description: |
    为AI编程大赛提供自动化的作品提交、评审任务调度、评分结果查询和排行榜服务。
    所有时间戳字段使用ISO 8601格式。
  version: 1.0.0
  contact:
    name: Hackathon Tech Team
servers:
  - url: http://localhost:8000/api/v1
    description: Local Development
  - url: https://hackathon2026.example.com/api/v1
    description: Production
paths:
...
  /submissions/{submission_id}/scores:
    get:
      summary: 获取作品的各维度评分
      operationId: getDimensionScores
      tags: [Scores]
      parameters:
        - name: submission_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: 各维度评分详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DimensionScores'
  /submissions/{submission_id}/report:
    get:
      summary: 获取评审报告（Markdown或PDF）
      operationId: getReviewReport
      tags: [Scores]
      parameters:
        - name: submission_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: format
          in: query
          schema:
            type: string
            enum: [markdown, pdf, json]
            default: markdown
      responses:
        '200':
          description: 评审报告
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewReport'
            text/markdown:
              schema: { type: string }
            application/pdf:
              schema: { type: string, format: binary }
  /submissions/{submission_id}/verification:
    get:
      summary: 获取Verifier交叉验证结果
      operationId: getVerificationResults
      tags: [Scores]
      parameters:
        - name: submission_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: 验证结果
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VerificationResult'

  # ============================================================
  # 排行榜
  # ============================================================
  /leaderboard:
    get:
      summary: 获取赛题排行榜
      operationId: getLeaderboard
      tags: [Leaderboard]
      parameters:
        - name: question_id
          in: query
          schema: { type: integer }
          description: 不填则返回总分榜
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: 排行榜数据
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Leaderboard'

  # ============================================================
  # 人工复核接口（评委专用）
  # ============================================================
  /human-reviews/pending:
    get:
      summary: 获取待人工复核的作品列表
      operationId: listPendingHumanReviews
      tags: [Human Review]
      parameters:
        - name: question_id
          in: query
          schema: { type: integer }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
      responses:
        '200':
          description: 待复核列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedSubmissions'
  /human-reviews/{submission_id}:
    post:
      summary: 提交人工复核评分
      operationId: submitHumanReview
      tags: [Human Review]
      parameters:
        - name: submission_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HumanReviewInput'
      responses:
        '200':
          description: 复核提交成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanReviewResult'
  /human-reviews/{submission_id}/materials:
    get:
      summary: 获取评委复核工作台所需的所有材料
      operationId: getReviewMaterials
      tags: [Human Review]
      parameters:
        - name: submission_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: 复核材料包
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewMaterials'

  # ============================================================
  # 系统监控与管理
  # ============================================================
...
components:
  schemas:
...
    SubmissionResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        team_id: { type: string, format: uuid }
        question_id: { type: integer }
        submission_version: { type: integer }
        file_path: { type: string }
        file_size: { type: integer }
        status:
          type: string
          enum: [pending, running, completed, failed]
        submitted_at: { type: string, format: date-time }
        review_task_id: { type: string, format: uuid, nullable: true }
        estimated_completion: { type: string, format: date-time, nullable: true }
    PaginatedSubmissions:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        page_size: { type: integer }
        items:
          type: array
          items:
            $ref: '#/components/schemas/SubmissionResponse'
...
    DimensionScores:
      type: object
      properties:
        submission_id: { type: string, format: uuid }
        scores:
          type: array
          items:
            type: object
            properties:
              dimension: { type: string }
              score: { type: integer, minimum: 0, maximum: 100 }
              confidence: { type: number, minimum: 0, maximum: 1 }
              reasoning: { type: string }
              strengths: { type: array, items: { type: string } }
              weaknesses: { type: array, items: { type: string } }
              improvements: { type: array, items: { type: string } }
              llm_model: { type: string }
              scored_at: { type: string, format: date-time }
        total_score: { type: number }
        overall_confidence: { type: number }
    VerificationResult:
      type: object
      properties:
        id: { type: string, format: uuid }
        verifier_model: { type: string }
        target_dimension: { type: string }
        original_score: { type: integer }
        verified_score: { type: integer }
        deviation: { type: number }
        confidence: { type: number }
        needs_human_review: { type: boolean }
        verification_reason: { type: string }
        created_at: { type: string, format: date-time }
    ReviewReport:
      type: object
      properties:
        submission_id: { type: string, format: uuid }
        status: { type: string, enum: [pending, generating, completed] }
        content: { type: string, description: "Markdown格式报告" }
        download_url: { type: string, format: uri }
        generated_at: { type: string, format: date-time }
    HumanReviewInput:
      type: object
      required: [reviewer_name, adjusted_scores]
      properties:
        reviewer_name: { type: string }
        adjusted_scores:
          type: object
          additionalProperties: { type: integer, minimum: 0, maximum: 100 }
          example:
            functionality: 85
            code_quality: 78
        adjustment_reason: { type: string }
        reviewer_notes: { type: string }
    HumanReviewResult:
      type: object
      properties:
        submission_id: { type: string, format: uuid }
        reviewer_name: { type: string }
        original_total: { type: number }
        adjusted_total: { type: number }
        finalized: { type: boolean }
        reviewed_at: { type: string, format: date-time }
    ReviewMaterials:
      type: object
      properties:
        submission_id: { type: string, format: uuid }
        project_files:
          type: array
          items:
            type: object
            properties:
              path: { type: string }
              size: { type: integer }
              language: { type: string }
        readme_content: { type: string }
        test_results: { type: object }
        agent_scores:
          type: array
          items:
            $ref: '#/components/schemas/DimensionScores/properties/scores/items'
        verification_results:
          type: array
          items:
            $ref: '#/components/schemas/VerificationResult'
        sandbox_logs:
          type: array
          items:
            type: object
            properties:
              execution_type: { type: string }
              exit_code: { type: integer }
              stdout: { type: string }
              stderr: { type: string }
        code_preview:
          type: array
          items:
            type: object
            properties:
              path: { type: string }
              content: { type: string }
              language: { type: string }
    Leaderboard:
      type: object
      properties:
        question_id: { type: integer, nullable: true }
        question_title: { type: string, nullable: true }
        total: { type: integer }
        page: { type: integer }
        page_size: { type: integer }
        items:
          type: array
          items:
            type: object
            properties:
              rank: { type: integer }
              team_id: { type: string, format: uuid }
              team_code: { type: string }
              team_name: { type: string }
              total_score: { type: number }
              confidence: { type: number }
              human_reviewed: { type: boolean }
              scored_at: { type: string, format: date-time }
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []